requirements.txt
kafka-python==2.0.2
clickhouse-connect==0.8.18
python-dotenv==1.0.1
pyspark==3.5.1
pytest==8.3.2
